requirements.txt
absl-py==0.7.1
affine==2.2.2
astor==0.8.0
attrs==19.1.0
boto3==1.9.191
botocore==1.12.191
catalyst==19.5
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cloudpickle==1.2.1
configparser==3.7.4
crc32c==1.7
cycler==0.10.0
decorator==4.4.0
Django==2.2.3
djangorestframework==3.10.1
docutils==0.15
Fiona==1.8.6
Flask==1.1.1
future==0.17.1
gast==0.2.2
geopandas==0.5.1
grpcio==1.22.0
gym==0.13.1
h5py==2.9.0
idna==2.8
imageio==2.5.0
invoke==1.2.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jpeg4py==0.1.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lz4==2.1.10
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mock==3.0.5
munch==2.3.2
networkx==2.3
numpy==1.16.4
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.1.0
plotly==4.0.0
pretrainedmodels==0.7.4
protobuf==3.9.0
psycopg2-binary==2.8.3
pyarrow==0.14.0
pycparser==2.19
pyglet==1.3.2
pymongo==3.8.0
pyparsing==2.4.0
pyproj==2.2.1
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.1
pyzmq==18.0.2
rasterio==1.0.24
redis==3.2.1
requests==2.22.0
retrying==1.3.3
s3transfer==0.2.1
safitty==1.2.3
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
segmentation-models-pytorch==0.0.1
sentinelhub==2.5.3
Shapely==1.6.4.post2
six==1.12.0
snuggs==1.4.6
sqlparse==0.3.0
tensorboard==1.13.1
tensorboardX==1.8
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tifffile==2019.7.2
torch==1.1.0
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.2.2
tornado==6.0.3
tqdm==4.19.9
urllib3==1.25.3
utm==0.5.0
visdom==0.1.8.8
websocket-client==0.56.0
Werkzeug==0.15.5
